refactor(sse): move the thinking-budget helpers out of base.ts
base.ts sits 44 lines over its frozen 1578-line ceiling, so the file-size gate
fails on it. The thinking-budget helpers are the natural piece to lift out: they
are pure functions over a request body, they touch nothing on the executor, and
they are the one self-contained cluster in the file.

hasActiveClaudeThinking, collectThinkingConfigs, readNestedThinkingBudget and
clampNestedThinkingBudget move verbatim to open-sse/utils/thinkingBudget.ts.
The bodies are byte-identical to the ones they replace and base.ts calls the
three it uses at the same four sites as before, so there is no behaviour change
to review -- only a boundary.

base.ts goes from 1622 to 1562 lines, sixteen under the ceiling, so the frozen
baseline does not need adjusting. The existing coverage in
tests/unit/base-thinking-budget-config-5312.test.ts still exercises the helpers
through base.ts and passes unchanged.
